+++ src/main/java/sa/gov/nic/bio/bw/client/core/CoreFxController.java
@@ -105,6 +105,15 @@ public class CoreFxController implements IdleMonitorRegisterer
 		
 		primaryStage.setTitle(windowTitle);
 		idleMonitor = new IdleMonitor(idleWarningBeforeSeconds, this::onShowingIdleWarning, idleWarningAfterSeconds, this::onIdle, this::onIdleInterrupt, this::onTick, idleNotifier);
+		
+		// workaround to bring the primary stage to the front in case the application has lost the focus on startup.
+		Platform.runLater(() ->
+		{
+			primaryStage.setAlwaysOnTop(true);
+			primaryStage.setAlwaysOnTop(false);
+			primaryStage.setIconified(true);
+			primaryStage.setIconified(false);
+		});
 	}
 	
 	@Override