+++ src/main/java/sa/gov/nic/bio/bw/client/core/utils/RuntimeEnvironment.java
@@ -1,0 +1,16 @@
+package sa.gov.nic.bio.bw.client.core.utils;
+
+public enum RuntimeEnvironment
+{
+	PROD, INT, DEV;
+	
+	public static RuntimeEnvironment byName(String name)
+	{
+		for(RuntimeEnvironment env : values())
+		{
+			if(env.name().equalsIgnoreCase(name)) return env;
+		}
+		
+		return null;
+	}
+}