+++ src/main/java/sa/gov/nic/bio/bw/client/login/ChangePasswordDialogController.java
@@ -1,0 +1,4 @@
+package sa.gov.nic.bio.bw.client.login;
+
+public class ChangePasswordDialogController
+{}