+++ src/main/java/sa/gov/nic/bio/bw/client/login/LoginPaneFxController.java
@@ -1,13 +1,15 @@
 package sa.gov.nic.bio.bw.client.login;
 
 import javafx.beans.binding.BooleanBinding;
-import javafx.beans.value.ObservableValue;
 import javafx.event.ActionEvent;
 import javafx.fxml.FXML;
+import javafx.geometry.NodeOrientation;
 import javafx.scene.control.*;
+import javafx.stage.Stage;
 import sa.gov.nic.bio.bw.client.core.*;
 import sa.gov.nic.bio.bw.client.core.beans.StateBundle;
 import sa.gov.nic.bio.bw.client.core.interfaces.*;
+import sa.gov.nic.bio.bw.client.core.utils.DialogUtils;
 import sa.gov.nic.bio.bw.client.core.utils.GuiLanguage;
 
 import java.util.HashMap;
@@ -20,7 +22,9 @@ import java.util.logging.Logger;
  */
 public class LoginPaneFxController extends BodyFxControllerBase implements LanguageSwitchingController
 {
+	private static final String FXML_CHANGE_PASSWORD = "sa/gov/nic/bio/bw/client/login/fxml/change_password_dialog.fxml";
 	private static final Logger LOGGER = Logger.getLogger(LoginPaneFxController.class.getName());
+	
 	@FXML private TextField txtUsername;
 	@FXML private PasswordField txtPassword;
 	@FXML private ComboBox<GuiLanguage> cbLanguage;
@@ -44,7 +48,11 @@ public class LoginPaneFxController extends BodyFxControllerBase implements Langu
 	{
 		GuiLanguage currentLanguage = coreFxController.getGuiState().getLanguage();
 		cbLanguage.getSelectionModel().select(currentLanguage);
-		cbLanguage.valueProperty().addListener(this::onSwitchingLanguage);
+		cbLanguage.valueProperty().addListener((observable, oldValue, newValue) ->
+		{
+			Locale.setDefault(newValue.getLocale());
+			coreFxController.switchLanguage(newValue, this);
+		});
 		
 		coreFxController.getMenuPaneController().hideRootPane();
 		coreFxController.getHeaderPaneController().hideRootPane();
@@ -63,12 +71,6 @@ public class LoginPaneFxController extends BodyFxControllerBase implements Langu
 		disableUiControls(false);
 	}
 	
-	private void onSwitchingLanguage(ObservableValue<? extends GuiLanguage> observable, GuiLanguage oldValue, GuiLanguage newValue)
-	{
-		Locale.setDefault(newValue.getLocale());
-		coreFxController.switchLanguage(newValue, this);
-	}
-	
 	@FXML
 	private void onLoginButtonClicked(ActionEvent event)
 	{
@@ -85,6 +87,19 @@ public class LoginPaneFxController extends BodyFxControllerBase implements Langu
 		coreFxController.submitFormTask(uiDataMap);
 	}
 	
+	@FXML
+	private void onChangePasswordButtonClicked(ActionEvent event)
+	{
+		notificationPane.hide();
+		
+		// TODO: show dialog
+		//String title = labelsBundle.getString("dialog.compare.title");
+		boolean rtl = coreFxController.getGuiState().getLanguage().getNodeOrientation() == NodeOrientation.RIGHT_TO_LEFT;
+		
+		Dialog<Void> dialogStage = DialogUtils.buildCustomDialog(appIcon, FXML_CHANGE_PASSWORD, labelsBundle, rtl);
+		dialogStage.showAndWait();
+	}
+	
 	private void disableUiControls(boolean bool)
 	{
 		txtUsername.setDisable(bool);
@@ -97,8 +112,8 @@ public class LoginPaneFxController extends BodyFxControllerBase implements Langu
 		btnLogin.setManaged(!bool);
 		btnLogin.setVisible(!bool);
 		
-		//btnChangePassword.setManaged(!bool);
-		//btnChangePassword.setVisible(!bool);
+		btnChangePassword.setManaged(!bool);
+		btnChangePassword.setVisible(!bool);
 	}
 	
 	@FXML